Added cannon mecanic :100: :100:

+++ TheAmazingCarAdventures/ModulePlayer.h
@@ -38,6 +38,6 @@ public:
 
 private:
 	btRigidBody* chassis_vehicle;
-	Uint32 best_time = 0;
+	float best_time = 0;
 	mat4x4 initial_car_matrix;
 };
